Remove functionality to output source files
The commit removes the code that generates source files with line
numbers.  JSDoc has an option to output these, and include links
in the documentation to the relevant lines in the doc. This adds
unnecessary complexity; if we link to source code, we should link to
the code in github instead.