Output documentation as a single erb file
This commit changes the documentation so that instead of being output
as individual HTML pages, it's instead output as a single block of
HTML, saved in _jsdoc.erb.  The resulting documentation is not
proper HTML because it's missing the necessary header; _jsdoc.erb
is intended to be inserted into another HTML file (Quilt uses
Middleman to insert this file into the rest of the Quilt
documentation).