Clean up the HTML for insertion into Quilt docs
This commit removes a bunch of HTML tags that create unnecessary
structure (and, as a result, lead to formatting problems when
the resulting HTML is inserted into the Quilt documentation). This
commit removes some elements, changes some <div> elements to paragraphs,
and uses smaller header sizes in a few places, for consistency with the
headers used in the rest of the Quilt documentation.