Fix links to point to IDs
The previous commit changed the docs generation to output one
block of HTML rather than individual pages. This broke the
links (e.g., the link to a type described in the docs), because
those links point at the pages that no longer exists.

This commit updates the documentation to use an ID at the top of each
major element (e.g., a class), and changes the links used in the
documentation to point at that id.