params: Don't nest the optional parameter table
We often use an optional parameter object in our constructors. In our
documentation, the possible properties of these objects would appear
as a nested parameter table as the "Description" of the optional object.
This looked terrible and was hard to read, so this commit gets rid
of the nested tables and instead lists parameters as opts.x and opts.y.