signature: Show optional parameters in brackets
Before, the template would append the string "opt" to all optional parameters
when showing the function signature. E.g. `optionalArgs` would become
`optionalArgsopt`. This is not very readable, so we now instead list the optional
parameters in square brackets -- e.g. `foo(bar[, baz][, qux])`. This is the
conventional way of doing it -- e.g. in the official Node.js documentation.